Fix parser bug: classes ended program
A class declaration class {} would not correctly process the closing
brace, causing any further lines of code to be ignored and the program
to be missing components. With this commit,

    class foo {
      method asString { "hello world" }
    }
    print(foo)

now produces output, while previously the AST would end prior to the
print request, resulting in a program that created the class method and
then did nothing.