Add circular interpreter in Grace-in-Grace
The ast module can now be mocked in the same way that the Java version
does, providing the bare minimum wrapping to create the in-memory AST
from the Grace-in-Grace parser's ast.* requests. The resulting AST can
then be evaluated as usual, allowing code to be executed from textual
source code entirely within Grace code.