Add preliminary CPS interpreter in Java
This version avoids some recursion-depth issues and allows experimenting
with different kinds of evaluation. In preliminary testing it shows
signs of being faster than the baseline version despite the extra
indirection, but the reason is unclear.

It is capable of running the parser identically to the main Java
version, but only supports hard-coded programs currently.